Modernize EuroTest Unit Test
Remove `assertExactFloat`, we can now use `assertSame`. PHPUnit has
deprecated and removed the 4th argument (delta) for a while now, so the
test did no longer what it was supposed to do.

Call `expectNotToPerformAssertions` in a test that tests the absence of
exception-throwing instead of asserting that true is true.

Explain edge case test value with comment